% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgraph.R
\name{kgraph}
\alias{kgraph}
\title{kgraph}
\usage{
kgraph(keyword = "", token = "", ids = "", hl = "", types = "",
  prefix = FALSE, limit = 10)
}
\arguments{
\item{keyword}{A character vector with the actual Google Knowledge Graph
Search query keyword. Note that only one keyword is allowed a time.}

\item{token}{A character string containing your Google API access token.}

\item{ids}{A vector of entity ID(s) to obtain the Knowledge Graph details.
Provide in the form of \dQuote{/m/062s4}.}

\item{hl}{A string specifying the ISO 639 language code (e.g. \dQuote{en} or
\dQuote{fr}). Default is English.}

\item{types}{A vector of character strings which restricts the returned
entities. For example, you can specify Person to restrict the results to
entities representing people. If multiple types are specified, returned
entities will contain one or more of these types. Full list of schemas
types is availabe (\url{http://schema.org/docs/full.html})}

\item{prefix}{If set to \dQuote{TRUE} prefix (initial substring) match
against names and aliases of entities is allowed. For example, a prefix
Jung will match entities and aliases such as Jung, Jungle, and Jung-ho
Kang. Default is \dQuote{FALSE}.}

\item{limit}{Sets the maximum of returned entities per call. Default is
\dQuote{10}, maximum allowed is \dQuote{20}}
}
\value{
Returns an object of class \dQuote{kgraph}. This is a list containing
  the entities returned from the call sorted in an ascending order relative
  to the relevance score.
}
\description{
Obtains Google Knowledge Graph Search entities for a given keyword or set of
entitiy ID's.
}
\details{
\code{kgraph} can be used to obtain Google Knowledge Graph Search entities
(\url{https://www.google.com/intl/es419/insidesearch/features/search/knowledge.html})
for any given keyword. The obtained \dQuote{kg-id} can then be used to obtain
Google Trends information based on Google's Topic Search, e.g. with the
package gtrendsR. This is a set of combined search queries for any language.
For example the topic search query for \dQuote{London the capital of
England}, will not only cover searches for the keyword \dQuote{London} but
also for \dQuote{Londres} in Spanish. Details between the differences of
search over terms and topics are highlighted here
\url{https://support.google.com/trends/answer/4359550}.
}
\note{
When using \code{types} often \dQuote{Error:400} is returned since not
  all schemas are available, i.e. \dQuote{Vehicle}. In this case it is
  advisable to play around with the types categories to find out the ones
  working.

  Provided token has limited quota. Please create and use your own token for
  large queries. Details on how to obtain your access token can be found here
  (\url{https://developers.google.com/knowledge-graph/how-tos/authorizing}).
}
\examples{
\donttest{
kg <- kgraph("Myst", types = "VideoGame")
# get google trends for the first entity using gtrendsR package
topicsearch <- gtrends(kgs$entities[[1]]$id, time = "all")
}

}
\author{
Oliver Schaer, \email{info@oliverschaer.ch}
}
